pyproject.toml
[project]
name = "spotifytolfm-py"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.8"
dependencies = [
    "pylast>=6.0.0",
    "python-dotenv>=1.0.0",
    "spotipy>=2.25.1",
]
